Tweak time range label positions.
Instead of centering a label within the available (clipped to screen)
rectangle, we now always center withing the actual (full) time range,
and then clamp to the screen. The makes it a bit more clear that a box
is large and goes off screen, and also reduces excessive label
animations.