Moving from GSS to LESS. This is paying back some technical debt and making life way better going forward.
LESS is a much better syntax and (in future commits) will help
clean up a lot of hardcoded/assumed values that were required due to the
lack of dependency ordering in GSS. Future commits will remove the 'wtf'
prefix from class names, use more LESS features (nesting/&/etc), etc.

The HUD styles are now split out from their templates and can be shared
with the application. The GSS compiler is used on the resulting LESS output
to add a renaming prefix that prevents collisions with on-page content.

Future commits will move the SettingsDialog from hud/ and into ui/, letting
us build a settings dialog for the app UI.

SVG icons have been moved out of the CSS and into embedded rules for the
HUD - still no need for extra files, but a much better build process.
This has not been setup for the app yet as the toolbar there is going to
be reworked. By removing the need for the hacky content:url() stuff the
icons now appear in IE9 and Opera.