Added a previous draw call button.
Added a button that seeks to the previous draw call within a step. If no draw call precedes the current call (or no call within the current step has run), then clicking on the button just seeks to the start of a step. This button nicely complements the next draw call button.