Polished graphics replay UI.
The event navigator was polished. Colors were changed to match Kennedy styles. A column for context handles was made. A bug in which contexts sometimes appeared in different orders after seeking backwards was fixed. The dialog for changing arguments has been altered to adjust its size based on the sizes of the inputs. The edit arguments button has been changed to a pencil icon.