Fixed context mismatch in graphics replay.
Fixed the graphics replay bug in which context handles in the event navigator sometimes did not match the contexts being acted upon. Changed ContextPool to create hashes of contexts not just based on type and attributes, but also based on canvas dimenions. Changed the CanvasArea UI component to not always remove all canvases during a backwards seek.