Fixed context size not updating bug.
Fixed a bug in which the size of the canvas was not reset when the playback was reset. This bug is noticeable when a playback calls setContext more than once to change the size of a canvas on a single context: The canvas would erringly not revert to the size at which the context was first created.