Build fixes for iOS and Android
Replace strchrnul usage with strchr (and avoid implicit conversion from name_spec_ to intermediate std::string value).

Adopt <chrono> for timing. (Note that steady_clock is not guaranteed to provide nanosecond resolution, but it does on Linux/Android and iOS, at least. And high_resolution_clock is not guaranteed to be monotonic.)